Remove Client and Admin models. Merge it into User.
